% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_american.r
\docType{data}
\name{all_american}
\alias{all_american}
\title{All American Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3277 rows and 8 columns.
}
\usage{
all_american
}
\description{
All American Data
}
\examples{
  all_american
}
\keyword{datasets}
