% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowls.r
\docType{data}
\name{bowls}
\alias{bowls}
\title{Bowls Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2474 rows and 10 columns.
}
\usage{
bowls
}
\description{
Bowls Data
}
\examples{
  bowls
}
\keyword{datasets}
