% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coaches.r
\docType{data}
\name{coaches}
\alias{coaches}
\title{Coaches Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 30466 rows and 10 columns.
}
\usage{
coaches
}
\description{
Coaches Data
}
\examples{
  coaches
}
\keyword{datasets}
