% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conference.r
\docType{data}
\name{conference}
\alias{conference}
\title{Conference Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 13587 rows and 3 columns.
}
\usage{
conference
}
\description{
Conference Data
}
\examples{
  conference
}
\keyword{datasets}
