% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn.r
\docType{data}
\name{espn}
\alias{espn}
\title{ESPN Recruiting Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1279 rows and 15 columns.
}
\usage{
espn
}
\description{
ESPN Recruiting Data
}
\examples{
  espn
}
\keyword{datasets}
