% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polls.r
\docType{data}
\name{polls}
\alias{polls}
\title{Polls Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1693 rows and 7 columns.
}
\usage{
polls
}
\description{
Polls Data
}
\examples{
  polls
}
\keyword{datasets}
