% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings.r
\docType{data}
\name{rankings}
\alias{rankings}
\title{Rankings Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4361 rows and 9 columns.
}
\usage{
rankings
}
\description{
Rankings Data
}
\examples{
  rankings
}
\keyword{datasets}
