% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rivals.r
\docType{data}
\name{rivals}
\alias{rivals}
\title{Rivals Recruiting Ranking Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1558 rows and 9 columns.
}
\usage{
rivals
}
\description{
Rivals Recruiting Ranking Data
}
\examples{
  rivals
}
\keyword{datasets}
