% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sos.r
\docType{data}
\name{sos}
\alias{sos}
\title{Strength of Schedule Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3875 rows and 13 columns.
}
\usage{
sos
}
\description{
Strength of Schedule Data
}
\examples{
  sos
}
\keyword{datasets}
