% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_names.r
\docType{data}
\name{team_names}
\alias{team_names}
\title{Team Name Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 125 rows and 5 columns.
}
\usage{
team_names
}
\description{
Team Name Data
}
\examples{
  team_names
}
\keyword{datasets}
