% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teams.r
\docType{data}
\name{teams}
\alias{teams}
\title{Teams Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 252 rows and 12 columns.
}
\usage{
teams
}
\description{
Teams Data
}
\examples{
  teams
}
\keyword{datasets}
