% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yby.r
\docType{data}
\name{yby}
\alias{yby}
\title{Year by Year Game Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 128155 rows and 10 columns.
}
\usage{
yby
}
\description{
Year by Year Game Data
}
\examples{
  yby
}
\keyword{datasets}
